Update CCNode.m
Fixes the case when an action runs on a node and the node "loses" its properties before the action was run. 

Example: A node is child of a layer that is scaled (CCScaleTo action), and the node doesn't scale. Same for CCMoveTo applied to the parent layer, the child node stays in the same place. This affects only the nodes that previously ran another action (even if the action was stopped).